backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
greenlet>=3.0.3
supabase>=2.3.4

# dYdX v4 - Using Indexer API directly (no SDK needed)

# Notifications
python-telegram-bot>=20.7
discord.py>=2.3.2
aiohttp>=3.9.1
requests>=2.31.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
cryptography>=42.0.0
PyJWT>=2.8.0

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
httpx>=0.25.0
tenacity>=8.2.3
rich>=13.7.1

# Logging & Monitoring
structlog>=24.1.0
python-json-logger>=2.0.7
